Move solvers to individual subfolders
