+++ newton/solvers/euler/particles.py
@@ -15,6 +15,7 @@
 
 import warp as wp
 
+import newton
 from newton.core import PARTICLE_FLAG_ACTIVE
 
 
@@ -98,7 +99,7 @@ def eval_particle_forces_kernel(
     particle_f[i] = f
 
 
-def eval_particle_forces(model, state, forces):
+def eval_particle_forces(model: newton.Model, state: newton.State, forces: wp.array(dtype=wp.vec3)):
     if model.particle_count > 1 and model.particle_max_radius > 0.0:
         wp.launch(
             kernel=eval_particle_forces_kernel,
