+++ newton/solvers/featherstone/__init__.py
@@ -13,18 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .euler import SemiImplicitSolver
-from .featherstone import FeatherstoneSolver
-from .mujoco import MuJoCoSolver
-from .solver import SolverBase
-from .vbd import VBDSolver
-from .xpbd import XPBDSolver
+from .solver_featherstone import FeatherstoneSolver
 
 __all__ = [
     "FeatherstoneSolver",
-    "MuJoCoSolver",
-    "SemiImplicitSolver",
-    "SolverBase",
-    "VBDSolver",
-    "XPBDSolver",
-]
+]
